+++ forui/test/src/widgets/calendar/grid_calendar_test.dart
@@ -440,10 +440,7 @@ void main() {
         TestScaffold.app(
           child: StatefulBuilder(
             builder: (context, setState) => FCalendar.grid(
-              selectionControl: .lifted(
-                selected: (date) => date == selected,
-                select: (date) => setState(() => selected = date),
-              ),
+              selectionControl: .liftedSingle(value: selected, onChange: (date) => setState(() => selected = date)),
               control: control(),
             ),
           ),